backend/requirements.txt
# =========================
# CORE FASTAPI
# =========================
fastapi==0.110.0
uvicorn==0.27.1

pydantic==2.6.1
pydantic-settings==2.1.0

python-multipart==0.0.9
aiofiles==23.2.1

# =========================
# DATABASE
# =========================
motor==3.3.2
pymongo==4.6.3
qdrant-client==1.7.3

# =========================
# AUTH SECURITY (FIXED)
# =========================
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4

# =========================
# ML / CV
# =========================
numpy==1.26.4

opencv-python==4.9.0.80
onnxruntime==1.17.1
insightface==0.7.3

# =========================
# TESTING
# =========================
pytest==8.0.0
httpx==0.27.0
pytest-asyncio==0.23.5

# =========================
# BUILD TOOLS
# =========================
setuptools>=65.0.0
wheel>=0.40.0